Move test to Tests folder
